requirements.txt
numpy~=1.26.3
scikit-learn~=1.2.2
tensorflow~=2.10.0
flask~=2.2.5
Pillow~=9.3.0
opencv-python-headless~=4.7.0.68
joblib~=1.2.0
